added image of layout and I found a bug in my logic. Upon entering a sector there may be more than just 1 blind turnout. Therefor a system is needed to store a variable amount. Recommend to fix a certain max amount due to fixed storage positions.
The transitioning from one section to another also come with a flaw. If the detector lies on a blind turnout, you could enter the wrong section. At first I was using the assumption that you can transition to 2 sectors depending on the state of a turnout. But it may very well be that there is just one other section to which you can transition